[Refactor] Reply Context에서 전역에서 사용될 상태만 가지도록 리팩토링 (2023/09/20)
