+++ src/contexts/Reply.jsx
@@ -1,4 +1,3 @@
-import pb from '@/api/pocketbase';
 import { node, string } from 'prop-types';
 import { createContext, useState } from 'react';
 
@@ -6,30 +5,6 @@ export const ReplyContext = createContext({});
 
 function ReplyProvider({ displayName = 'ReplyContext', children }) {
   const [openModal, setOpenModal] = useState(false);
-  const [selectedComment, setSelectedComment] = useState({});
-
-  const handleOpenModal = (e, commentId) => {
-    if (openModal === false) {
-      if (e.key === 'Enter' || e.type === 'click') {
-        setOpenModal(true);
-      }
-      scrollTo({ top: 100000, behavior: 'smooth' });
-
-      (async () => {
-        const commentData = await pb.collection('comments').getOne(commentId, {
-          expand: 'commenter, reply',
-        });
-
-        setSelectedComment({
-          id: commentData.id,
-          nickname: commentData.expand.commenter.nickname,
-          reply: commentData.reply,
-        });
-      })();
-
-      return;
-    }
-  };
 
   return (
     <ReplyContext.Provider
@@ -37,8 +12,6 @@ function ReplyProvider({ displayName = 'ReplyContext', children }) {
       value={{
         openModal,
         setOpenModal,
-        handleOpenModal,
-        selectedComment,
       }}
     >
       {children}
